+++ packages/smarthr-ui/src/components/Scroller/client/index.ts
@@ -1,0 +1,1 @@
+export { Scroller } from './Scroller'
